Внесла исправления
