+++ db.php
@@ -1,0 +1,7 @@
+<?php
+return $params = array (
+        'host' => 'localhost',
+        'user' => 'root',
+        'password' => '',
+        'db_name' => 'doingsdone'
+);
