+++ index.php
@@ -1,7 +1,9 @@
 <?php
 require_once 'helpers.php';
+require_once 'db.php';
 
-$con = mysqli_connect("localhost", "root", "", "doingsdone");
+$con = mysqli_connect($params['host'], $params['user'], $params['password'], $params['db_name']);
+
 mysqli_set_charset($con, 'utf8');
 
 if (!$con) {
@@ -40,6 +42,21 @@ function isValueInArray($array, $key, $value)
     }
     return false;
 }
+
+/*функция, возвращающая значение массива по ключу при его наличии*/
+function getValue ($array, $key)
+{
+    if (isset($array[$key])) {
+        return $array[$key];
+    }
+}
+
+/*функция, возвращающая url*/
+function getUrl ($file_path)
+{
+    return str_replace($_SERVER['DOCUMENT_ROOT'], 'http://'.$_SERVER['HTTP_HOST'], $file_path);
+}
+
 
 /*функция, возвращающая массив проектов для конкретного пользователя*/
 function getProjects($con, int $user_id)
@@ -86,164 +103,11 @@ function getTasks($con, int $user_id, int $project_id = null)
     return $tasks;
 }
 
-/*функция для добавления задачи в БД*/
-function addTask ($con, int $user_id, string $task_name, int $project_id, string $due_date, string $file_path) {
-
-    $parameters = [$user_id, $task_name, $project_id];
-    $sql = 'INSERT INTO tasks (user_id, name, project_id';
-
-    if (!empty($due_date)) {
-        $parameters[] = $due_date;
-        $sql .= ', due_date';
-    }
-    if (!empty($file_path)) {
-        $parameters[] = $file_path;
-        $sql .= ', file_path';
-    }
-
-    $sql .= ') VALUES (';
-    for ($i = 0; $i < count($parameters); $i++) {
-        $sql .= '?, ';
-    }
-    $sql = substr($sql, 0, -2).')';
-
-    $stmt = db_get_prepare_stmt($con, $sql, $parameters);
-    return mysqli_stmt_execute($stmt);
-}
-
-/*функция, возвращающая url*/
-function getUrl ($file_path)
-{
-    return str_replace($_SERVER['DOCUMENT_ROOT'], 'http:\\\\'.$_SERVER['HTTP_HOST'], $file_path);
-}
-
-/*функция, возвращающая значение поля формы*/
-function getPostVal($name)
-{
-    return $_POST[$name] ?? '';
-}
-
-/*функция, возвращающая имя файла из формы*/
-function getFilesVal($name)
-{
-    if (isset ($_FILES[$name])) {
-        return $_FILES[$name]['name'] ?? '';
-    }
-}
-
-/*функция, возвращающая значение массива по ключу при его наличии*/
-function getValue ($array, $key)
-{
-    if (isset($array[$key])) {
-        return $array[$key];
-    }
-}
-
-/*функция для проверки заполненности поля формы*/
-function validateFilled($name)
-{
-    if (empty($_POST[$name])) {
-        return 'Это поле должно быть заполнено';
-    }
-}
-
-/*функция для валидации проекта*/
-function validateRealProject($projects)
-{
-    if (!isValueInArray($projects, 'id', $_POST['project'])) {
-        return 'Проект должен быть реально существующим';
-    }
-}
-
-/*функция для валидации даты*/
-function validateDate()
-{
-    $date = $_POST['date'];
-    if (!empty($date)) {
-        if (is_date_valid($date)) {
-            $cur_date = time();
-            $task_date = strtotime($date);
-            if (floor(($cur_date - $task_date) / 3600) >= 24) {
-                return 'Дата должна быть больше или равна текущей';
-            }
-        } else {
-            return 'Дата должна быть в формате ГГГГ-ММ-ДД';
-        }
-    }
-}
-
-/*функция, возвращающая класс для поля с ошибкой*/
-function getClassError ($errors, $name)
-{
-    if (isset($errors[$name])){
-        return 'form__input--error';
-    }
-    return '';
-}
-
-/*функция для добавления атрибута выбранному селекту*/
-function getSelected ($name, $id)
-{
-   if(isset($_POST[$name]) && getPostVal($name) == $id) {
-        return 'selected';
-    }
-}
-
-/*функция, возвращающая массив ошибок*/
-function getErrors ($projects)
-{
-    $errors = [];
-
-    $rules = [
-        'name' => function () {
-            return validateFilled('name');
-        },
-
-        'project' => function($projects) {
-            return validateRealProject($projects);
-        },
-
-        'date' => function() {
-            return validateDate();
-        }
-    ];
-
-    foreach ($_POST as $key => $value) {
-
-        if (isset($rules[$key])) {
-            $rule = $rules[$key];
-            $errors[$key] = $rule($projects);
-        }
-    }
-
-    return array_filter($errors);
-}
-
-/*функция для обработки формы добавления задачи*/
-function processingFormAddTask ($con, $user_id, $errors)
-{
-    $task_name = getPostVal('name');
-    $project_id = getPostVal('project');
-    $due_date = getPostVal('date');
-    $file_name = getFilesVal('file');
-    $file_path = '';
-
-    if (!count($errors)) {
-        if (!empty($file_name) && isset($_FILES['file'])) {
-            move_uploaded_file($_FILES['file']['tmp_name'], $_SERVER['DOCUMENT_ROOT'].'\\'.$file_name);
-            $file_path = $_SERVER['DOCUMENT_ROOT'].'\\'.$file_name;
-        }
-        addTask($con, $user_id, $task_name, $project_id, $due_date, $file_path);
-        header('Location: index.php');
-    }
-}
-
 /*объявление переменных*/
 $user_id = 1;
 $project_id = null;
 $show_complete_tasks = rand(0, 1);
 $projects = getProjects($con, $user_id);
-$errors = getErrors($projects);
 
 /*проверка выбранного id проекта в адресной строке*/
 if (isset($_GET['project_id'])) {
@@ -258,16 +122,10 @@ if (isset($_GET['project_id'])) {
 $tasks = array_reverse(getTasks($con, $user_id, $project_id));
 $tasksAll = array_reverse(getTasksAll($con, $user_id));
 
-
-/*шаблоны*/
-$add_content = include_template('add.php', ['projects' => $projects, 'tasksAll' => $tasksAll, 'con' => $con, 'user_id' => $user_id, 'errors' => $errors]);
-
-$main_content = include_template('main.php', ['show_complete_tasks' => $show_complete_tasks, 'projects' => $projects, 'tasks' => $tasks, 'tasksAll' => $tasksAll]);
-
-$layout_content = include_template('layout.php', ['content' => $main_content, 'title' => 'Дела в порядке', 'user_name' => 'Константин']);
-
-if (isset($_GET['add'])) {
-    $layout_content = include_template('layout.php', ['content' => $add_content, 'title' => 'Дела в порядке', 'user_name' => 'Константин']);
+/*подключение шаблона*/
+if ($_SERVER['PHP_SELF'] == '/index.php') {
+    $main_content = include_template('main.php', ['show_complete_tasks' => $show_complete_tasks, 'projects' => $projects, 'tasks' => $tasks, 'tasksAll' => $tasksAll]);
+    $layout_content = include_template('layout.php', ['content' => $main_content, 'title' => 'Дела в порядке', 'user_name' => 'Константин']);
+    print($layout_content);
 }
 
-print($layout_content);
